+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/diagnostics/DiagnosticsHelper.java
@@ -3,14 +3,10 @@
 
 package com.microsoft.applicationinsights.agent.internal.diagnostics;
 
-import com.azure.monitor.opentelemetry.exporter.implementation.statsbeat.MetadataInstanceResponse;
 import com.azure.monitor.opentelemetry.exporter.implementation.statsbeat.RpAttachType;
-import com.azure.monitor.opentelemetry.exporter.implementation.utils.PropertyHelper;
 import com.azure.monitor.opentelemetry.exporter.implementation.utils.Strings;
-import com.azure.monitor.opentelemetry.exporter.implementation.utils.SystemInformation;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import org.slf4j.LoggerFactory;
 
 public class DiagnosticsHelper {
   private DiagnosticsHelper() {}
@@ -19,7 +15,6 @@ public class DiagnosticsHelper {
   public static volatile boolean appSvcRpIntegration;
   private static volatile boolean functionsRpIntegration;
 
-  private static volatile char rpIntegrationChar;
   private static final boolean isWindows;
 
   public static final String LINUX_DEFAULT = "/var/log/applicationinsights";
@@ -41,32 +36,22 @@ public class DiagnosticsHelper {
     // TODO (heya) how should we report functions windows users who are using app services
     //  windows attach by manually setting the env vars (which was the old documented way)
     if ("java".equals(System.getenv("FUNCTIONS_WORKER_RUNTIME"))) {
-      rpIntegrationChar = 'f';
+      SdkVersionPrefixHelper.setRpIntegrationChar('f');
       functionsRpIntegration = true;
       setRpAttachType(agentPath, "functions.codeless");
     } else if (!Strings.isNullOrEmpty(System.getenv("WEBSITE_SITE_NAME"))) {
-      rpIntegrationChar = 'a';
+      SdkVersionPrefixHelper.setRpIntegrationChar('a');
       appSvcRpIntegration = true;
       setRpAttachType(agentPath, "appsvc.codeless");
     } else if (!Strings.isNullOrEmpty(System.getenv("KUBERNETES_SERVICE_HOST"))) {
-      rpIntegrationChar = 'k';
+      SdkVersionPrefixHelper.setRpIntegrationChar('k');
       setRpAttachType(agentPath, "aks.codeless");
     } else if (!Strings.isNullOrEmpty(
         System.getenv("APPLICATIONINSIGHTS_SPRINGCLOUD_SERVICE_ID"))) {
-      rpIntegrationChar = 's';
+      SdkVersionPrefixHelper.setRpIntegrationChar('s');
       setRpAttachType(agentPath, "springcloud.codeless");
     }
     // TODO (heya) detect VM environment by checking the AzureMetadataService response, manual only
-  }
-
-  /** Is resource provider (Azure Spring Cloud, AppService, Azure Functions, AKS, VM...). */
-  public static boolean isRpIntegration() {
-    return rpIntegrationChar != 0;
-  }
-
-  // returns 0 if not rp integration
-  public static char rpIntegrationChar() {
-    return rpIntegrationChar;
   }
 
   public static boolean isAppSvcRpIntegration() {
@@ -75,12 +60,6 @@ public class DiagnosticsHelper {
 
   public static boolean isFunctionsRpIntegration() {
     return functionsRpIntegration;
-  }
-
-  public static void lazyUpdateVmRpIntegration(MetadataInstanceResponse response) {
-    rpIntegrationChar = 'v';
-    PropertyHelper.setSdkNamePrefix(getRpIntegrationSdkNamePrefix());
-    RpAttachType.setRpAttachType(RpAttachType.STANDALONE_AUTO);
   }
 
   public static ApplicationMetadataFactory getMetadataFactory() {
@@ -97,21 +76,5 @@ public class DiagnosticsHelper {
     } else {
       RpAttachType.setRpAttachType(RpAttachType.STANDALONE_AUTO);
     }
-  }
-
-  public static String getRpIntegrationSdkNamePrefix() {
-    StringBuilder sdkNamePrefix = new StringBuilder(3);
-    sdkNamePrefix.append(DiagnosticsHelper.rpIntegrationChar());
-    if (SystemInformation.isWindows()) {
-      sdkNamePrefix.append("w");
-    } else if (SystemInformation.isLinux()) {
-      sdkNamePrefix.append("l");
-    } else {
-      LoggerFactory.getLogger("com.microsoft.applicationinsights.agent")
-          .warn("could not detect os: {}", System.getProperty("os.name"));
-      sdkNamePrefix.append("u");
-    }
-    sdkNamePrefix.append("_");
-    return sdkNamePrefix.toString();
   }
 }
